timer update

+++ Marlin/HAL.h
@@ -99,7 +99,7 @@ unsigned char eeprom_read_byte(unsigned char *pos);
 #define TEMP_TIMER_NUM 3
 #define BEEPER_TIMER_NUM 4
 
-#define HAL_TIMER_RATE 		     (F_CPU/2.0)
+#define HAL_TIMER_RATE 		     (long)(F_CPU/2)
 #define TICKS_PER_NANOSECOND   (HAL_TIMER_RATE)/1000
 
 #define ENABLE_STEPPER_DRIVER_INTERRUPT()	HAL_timer_enable_interrupt (STEP_TIMER_NUM)
